Moved packet to Core.RoveProtocol
