+++ Core/Interfaces/IRovecomm.cs
@@ -1,10 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Net;
-using Core.Models;
+﻿using System.Net;
+using Core.RoveProtocol;
 
 namespace Core.Interfaces
 {
