+++ RED/ViewModels/Modules/CameraViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using Caliburn.Micro;
 using Core.Interfaces;
-using Core.Models;
+using Core.RoveProtocol;
 
 namespace RED.ViewModels.Modules
 {
