+++ RED/ViewModels/Modules/SensorViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using Caliburn.Micro;
 using Core.Interfaces;
-using Core.Models;
+using Core.RoveProtocol;
 using RED.Models.Modules;
 using RED.Models.Network;
 using System;
